Refactored code to emulate the architecture - Removed workflow
