+++ src/main/java/com/careerfair/q/controller/student/implementation/StudentControllerImpl.java
@@ -1,13 +1,13 @@
-package com.careerfair.q.student.controller.implementation;
+package com.careerfair.q.controller.student.implementation;
 
-import com.careerfair.q.student.controller.StudentController;
-import com.careerfair.q.student.service.StudentService;
-import com.careerfair.q.student.workflow.request.AddStudentRequest;
-import com.careerfair.q.student.workflow.request.UpdateStudentRequest;
-import com.careerfair.q.student.workflow.response.AddStudentResponse;
-import com.careerfair.q.student.workflow.response.DeleteStudentResponse;
-import com.careerfair.q.student.workflow.response.GetStudentResponse;
-import com.careerfair.q.student.workflow.response.UpdateStudentResponse;
+import com.careerfair.q.controller.student.StudentController;
+import com.careerfair.q.service.student.StudentService;
+import com.careerfair.q.service.student.request.AddStudentRequest;
+import com.careerfair.q.service.student.request.UpdateStudentRequest;
+import com.careerfair.q.service.student.response.AddStudentResponse;
+import com.careerfair.q.service.student.response.DeleteStudentResponse;
+import com.careerfair.q.service.student.response.GetStudentResponse;
+import com.careerfair.q.service.student.response.UpdateStudentResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
