--- src/main/java/com/careerfair/q/database/service/DatabaseService.java
@@ -1,4 +1,0 @@
-package com.careerfair.q.database.service;
-
-public interface DatabaseService {
-}
