+++ src/main/java/com/careerfair/q/service/student/StudentService.java
@@ -1,11 +1,11 @@
-package com.careerfair.q.student.service;
+package com.careerfair.q.service.student;
 
-import com.careerfair.q.student.workflow.request.AddStudentRequest;
-import com.careerfair.q.student.workflow.request.UpdateStudentRequest;
-import com.careerfair.q.student.workflow.response.AddStudentResponse;
-import com.careerfair.q.student.workflow.response.DeleteStudentResponse;
-import com.careerfair.q.student.workflow.response.GetStudentResponse;
-import com.careerfair.q.student.workflow.response.UpdateStudentResponse;
+import com.careerfair.q.service.student.request.AddStudentRequest;
+import com.careerfair.q.service.student.request.UpdateStudentRequest;
+import com.careerfair.q.service.student.response.AddStudentResponse;
+import com.careerfair.q.service.student.response.DeleteStudentResponse;
+import com.careerfair.q.service.student.response.GetStudentResponse;
+import com.careerfair.q.service.student.response.UpdateStudentResponse;
 
 public interface StudentService {
