+++ src/main/java/com/careerfair/q/service/student/implementation/StudentServiceImpl.java
@@ -1,18 +1,18 @@
-package com.careerfair.q.student.workflow.implementation;
+package com.careerfair.q.service.student.implementation;
 
-import com.careerfair.q.database.service.student.StudentFirebase;
-import com.careerfair.q.student.workflow.StudentWorkflow;
-import com.careerfair.q.student.workflow.request.AddStudentRequest;
-import com.careerfair.q.student.workflow.request.UpdateStudentRequest;
-import com.careerfair.q.student.workflow.response.AddStudentResponse;
-import com.careerfair.q.student.workflow.response.DeleteStudentResponse;
-import com.careerfair.q.student.workflow.response.GetStudentResponse;
-import com.careerfair.q.student.workflow.response.UpdateStudentResponse;
+import com.careerfair.q.service.database.StudentFirebase;
+import com.careerfair.q.service.student.StudentService;
+import com.careerfair.q.service.student.request.AddStudentRequest;
+import com.careerfair.q.service.student.request.UpdateStudentRequest;
+import com.careerfair.q.service.student.response.AddStudentResponse;
+import com.careerfair.q.service.student.response.DeleteStudentResponse;
+import com.careerfair.q.service.student.response.GetStudentResponse;
+import com.careerfair.q.service.student.response.UpdateStudentResponse;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Service;
 
-@Component
-public class StudentWorkflowImpl implements StudentWorkflow {
+@Service
+public class StudentServiceImpl implements StudentService {
 
     @Autowired private StudentFirebase studentFirebase;
