+++ src/main/java/com/careerfair/q/service/student/request/AddStudentRequest.java
@@ -1,0 +1,4 @@
+package com.careerfair.q.service.student.request;
+
+public class AddStudentRequest {
+}
