+++ src/main/java/com/careerfair/q/service/student/response/StudentResponse.java
@@ -1,0 +1,4 @@
+package com.careerfair.q.service.student.response;
+
+public abstract class StudentResponse {
+}
