--- src/main/java/com/careerfair/q/student/workflow/request/AddStudentRequest.java
@@ -1,4 +1,0 @@
-package com.careerfair.q.student.workflow.request;
-
-public class AddStudentRequest {
-}
