--- src/main/java/com/careerfair/q/student/workflow/response/StudentResponse.java
@@ -1,4 +1,0 @@
-package com.careerfair.q.student.workflow.response;
-
-public abstract class StudentResponse {
-}
